Split utils into a package
